Remove unused option methods.
